# HADOOP-19389: Optimize shell -text command I/O with multi-byte read.