@cnauroth 2025-01-17
> no comments on the main code, more on tests.
> 
> I'd like this new code to become a contract test, so we can make sure it goes ok through cloud stores and hdfs

Thanks, @steveloughran ! I entered a question about the contract test feedback. For all other feedback, I've updated the patch.