@cnauroth 2025-01-17
The remaining checkstyle issues getting flagged are "variable must be private and have accessor methods." These are on inner classes, and I think only getting flagged due to me doing some refactoring improvements, so I'm not planning on addressing these.