@cnauroth 2025-01-24
I merged this to trunk and branch-3.4. Thanks for the reviews, @ashutoshcipher and @steveloughran !

I filed [HADOOP-19400](https://issues.apache.org/jira/browse/HADOOP-19400) for follow-up on new specs/contract tests for stream read behavior.